moving vocabulary-related functions from helpers to vocabulary.py
